requirements.txt
aiosignal==1.3.1
amqp==5.3.1
appdirs==1.4.4
area==1.1.1
asgiref==3.9.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
babel==2.17.0
beautifulsoup4==4.12.2
billiard==4.2.1
black==25.1.0
bleach==6.0.0
celery==5.5.3
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssutils==2.7.0
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==5.2.4
django-admin-autocomplete-all==0.6.6
django-admin-json-editor==0.2.3
django-autoslug==1.9.9
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-currentuser==0.9.0
django-import-export==3.2.0
django-jazzmin==3.0.1
django-json-widget==2.0.1
django-pandas==0.6.6
django-pdfkit==0.3.1
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docxcompose==1.4.0
docxtpl==0.20.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
drf-yasg==1.21.5
et-xmlfile==1.1.0
fastjsonschema==2.17.1
fonttools==4.39.4
frozenlist==1.7.0
fs==2.4.16
GDAL==3.3.2
gprof2dot==2022.7.29
gunicorn==20.1.0
html2image==2.0.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
kombu==5.5.4
lxml==6.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.10.3
mistune==2.0.5
multidict==6.0.4
mypy_extensions==1.1.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
numpy==2.3.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.3.1
pandocfilters==1.5.0
pathspec==0.12.1
pdfkit==0.6.0
pillow==11.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
prompt-toolkit==3.0.38
propcache==0.3.2
psycopg2-binary==2.9.10
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyshp==2.3.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.2.0
pytz==2023.3
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tablib==3.4.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.14.1
tzdata==2025.2
unicode-slugify==0.1.5
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.6
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.20.1
zipp==3.15.0
